Improved Vulkan Performance
- Added Single Write option to RenderPassSpecification
- Added Framebuffer PrepareImages
- Framebuffer Clear commands are grouped into a single command buffer
- Removed GraphicsContext GetImageFormat
- Removed Renderer2D BindRenderPass
- Improved Debug Profiling frame stats

+++ Nebula-Storm/src/Panels/Content_Browser.cpp
@@ -490,7 +490,7 @@ namespace Nebula {
 			{
 				currentPath /= p;
 
-				uint32_t index = m_TreeNodes.find(currentPath);
+				size_t index = m_TreeNodes.find(currentPath);
 				if (index >= m_TreeNodes.size())
 					m_TreeNodes.push_back(currentPath);
 			}
